AI_Recommend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
scikit-learn
python-multipart
